[feat] Authentication Layer
