+++ packages/nextjs/.eslintrc.json
@@ -1,3 +1,6 @@
 {
-  "extends": "next"
+  "extends": "next",
+  "rules": {
+    "react-hooks/exhaustive-deps": "off"
+  }
 }
