+++ packages/nextjs/hooks/scaffold-stark/index.ts
@@ -3,5 +3,4 @@ export * from "./useDeployedContractInfo";
 export * from "./useNetworkColor";
 export * from "./useAnimationConfig";
 export * from "./useTransactor";
-export * from "./useAutoConnect";
 export * from "./useSwitchNetwork";
