+++ packages/nextjs/hooks/scaffold-stark/useScaffoldEventHistory.ts
@@ -160,14 +160,12 @@ export const useScaffoldEventHistory = <
 
   useEffect(() => {
     readEvents(fromBlock).then();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [fromBlock, enabled]);
 
   useEffect(() => {
     if (!deployedContractLoading) {
       readEvents().then();
     }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [
     provider,
     contractName,
@@ -175,7 +173,6 @@ export const useScaffoldEventHistory = <
     deployedContractLoading,
     deployedContractData?.address,
     deployedContractData,
-    // eslint-disable-next-line react-hooks/exhaustive-deps
     JSON.stringify(filters, replacer),
     blockData,
     transactionData,
